Completed Application with DB

+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/AdminPrivilegesTable.swift
@@ -3,6 +3,6 @@ import SQLite
 enum AdminPrivilegesTable {
     static let table = Table("admin_privileges")
     
-    static let adminId = Expression<Int>("admin_id")
+    static let adminId = Expression<Int64>("admin_id")
     static let privilege = Expression<String>("privilege")
 }
