+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/BookingSeatsTable.swift
@@ -3,6 +3,6 @@ import SQLite
 enum BookingSeatsTable {
     static let table = Table("bookings_seats")
     
-    static let bookingId = Expression<Int>("booking_id")
-    static let showSeatId = Expression<Int>("show_seat_id")
+    static let bookingId = Expression<Int64>("booking_id")
+    static let showSeatId = Expression<Int64>("show_seat_id")
 }
