+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/CinemaHallsTable.swift
@@ -3,7 +3,7 @@ import SQLite
 enum CinemaHallsTable {
     static let table = Table("cinema_halls")
     
-    static let id = Expression<Int>("id")
+    static let id = Expression<Int64>("id")
     static let cinemaHallName = Expression<String>("name")
-    static let theatreId = Expression<Int>("theatre_id")
+    static let theatreId = Expression<Int64>("theatre_id")
 }
