+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/MovieLanguagesTable.swift
@@ -3,6 +3,6 @@ import SQLite
 enum MovieLanguagesTable {
     static let table = Table("movie_languages")
     
-    static let movieId = Expression<Int>("movie_id")
+    static let movieId = Expression<Int64>("movie_id")
     static let language = Expression<String>("language")
 }
