+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/SeatsTable.swift
@@ -3,9 +3,9 @@ import SQLite
 enum SeatsTable {
     static let table = Table("seats")
     
-    static let id = Expression<Int>("id")
+    static let id = Expression<Int64>("id")
     static let row = Expression<String>("row")
     static let seatNumber = Expression<Int>("seat_number")
     static let type = Expression<String>("type")
-    static let cinemaHallId = Expression<Int>("cinema_hall_id")
+    static let cinemaHallId = Expression<Int64>("cinema_hall_id")
 }
