+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/ShowSeatsTable.swift
@@ -3,8 +3,8 @@ import SQLite
 enum ShowSeatsTable {
     static let table = Table("show_seats")
     
-    static let id = Expression<Int>("id")
-    static let seatId = Expression<Int>("seat_id")
-    static let showId = Expression<Int>("show_id")
+    static let id = Expression<Int64>("id")
+    static let seatId = Expression<Int64>("seat_id")
+    static let showId = Expression<Int64>("show_id")
     static let status = Expression<String>("status")
 }
