+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/TheatresTable.swift
@@ -3,7 +3,7 @@ import SQLite
 enum TheatresTable {
     static let table = Table("theatres")
     
-    static let id = Expression<Int>("id")
+    static let id = Expression<Int64>("id")
     static let theatreName = Expression<String>("name")
     static let address = Expression<String>("address")
 }
