+++ MovieTicketBookingApplicationWithDB/Sources/Repositories/SQLiteImplementation/Tables/UsersTable.swift
@@ -3,7 +3,7 @@ import SQLite
 enum UsersTable {
     static let table = Table("users")
     
-    static let id = Expression<Int>("id")
+    static let id = Expression<Int64>("id")
     static let username = Expression<String>("username")
     static let password = Expression<String>("password")
     static let role = Expression<String>("role")
